Fixed bug where translated error message would not make sense (weird charactes would pop up), and fixed bunch of warning

+++ hw3D/Timer.cpp
@@ -17,5 +17,5 @@ float Timer::Reset()
 
 float Timer::Get() const
 {
-	return (steady_clock::now() - last).count();
+	return static_cast<float>( (steady_clock::now() - last).count() );
 }
